Fix a few bugs from upstream

+++ reimbursement/signals.py
@@ -8,8 +8,10 @@ from reimbursement.models import Reimbursement
 @receiver(post_save, sender=Application)
 def reimbursement_create(sender, instance, created, *args, **kwargs):
     exists = Reimbursement.objects.filter(hacker=instance.user).exists()
-    if not instance.reimb:
+    if not instance.reimb and exists:
         Reimbursement.objects.get(hacker=instance.user).delete()
+        return
+    if not instance.reimb:
         return
     if exists:
         reimb = Reimbursement.objects.get(hacker=instance.user)
